(#45): AroundContentTypeId를 '@/pages/types'에서 가져오도록 수정
