+++ src/lib/isValidTourType.ts
@@ -1,5 +1,5 @@
 import { TOUR_TYPE } from '@/pages/const/MARKER';
-import type { AroundContentTypeId } from '@/pages/map/aroundSearch/types';
+import type { AroundContentTypeId } from '@/pages/types';
 
 const isValidTourType = (value: string): value is AroundContentTypeId => {
   return value in TOUR_TYPE;
