+++ src/pages/tour/components/withGeoTripParams.tsx
@@ -1,7 +1,6 @@
 import { useSearchParams } from 'react-router-dom';
-import type { GeoTripLocation } from '@/pages/types';
+import type { AroundContentTypeId, GeoTripLocation } from '@/pages/types';
 import getSuspenseLocation from '@/lib/getSuspenseLocation';
-import type { AroundContentTypeId } from '@/pages/map/aroundSearch/types';
 import { isValidTourType } from '@/lib';
 
 interface InjectedProps {
@@ -16,15 +15,15 @@ export default function withGeoTripParams<P extends InjectedProps>(
   return function GeoTripWrapper(props: Omit<P, keyof InjectedProps>) {
     const [searchParams] = useSearchParams();
     const distance = searchParams.get('distance');
-    const tourType = searchParams.get('tour-type');
+    const tourContentTypeId = searchParams.get('tour-type');
     const geoLocation = getSuspenseLocation();
 
-    if (!distance || !tourType) {
+    if (!distance || !tourContentTypeId) {
       return (
         <div>필요한 정보가 부족합니다. 거리와 관광지 타입을 확인해주세요.</div>
       );
     }
-    if (!isValidTourType(tourType)) {
+    if (!isValidTourType(tourContentTypeId)) {
       return <div>잘못된 관광 타입입니다.</div>;
     }
 
@@ -33,7 +32,7 @@ export default function withGeoTripParams<P extends InjectedProps>(
         {...(props as P)}
         location={geoLocation}
         distance={distance}
-        tourType={tourType}
+        tourContentTypeId={tourContentTypeId}
       />
     );
   };
